+++ app/src/main/res/layout/fragment_home.xml
@@ -1,117 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
+<FrameLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
-             xmlns:tools="http://schemas.android.com/tools"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-             tools:context=".fragment.HomeFragment">
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".fragment.HomeFragment">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-                android:id="@+id/mainSwipeRefreshLayout"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-
-            <com.apachat.swipereveallayout.core.SwipeLayout
-                    android:id="@+id/quoteSwipeLayout"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    app:dragEdge="right"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    app:mode="same_level">
-
-                <!-- Creator Profile Layout -->
-                <FrameLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent">
-
-                    <LinearLayout
-                            android:id="@+id/includeAccountProfile"
-                            android:layout_width="match_parent"
-                            android:layout_height="match_parent"
-                            android:visibility="visible"
-                            app:layout_constraintBottom_toBottomOf="parent"
-                            app:layout_constraintEnd_toEndOf="parent"
-                            app:layout_constraintStart_toStartOf="parent"
-                            app:layout_constraintTop_toTopOf="parent">
-
-                        <include layout="@layout/user_account"/>
-                    </LinearLayout>
-
-                </FrameLayout>
-
-                <!-- RecyclerView -->
-                <FrameLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent">
-
-                    <androidx.recyclerview.widget.RecyclerView
-                            android:id="@+id/quoteRecyclerView"
-                            android:layout_width="match_parent"
-                            android:layout_height="match_parent"
-                            android:background="@color/transparent"
-                            app:layout_constraintBottom_toTopOf="@+id/createQuote"
-                            app:layout_constraintEnd_toEndOf="parent"
-                            app:layout_constraintHorizontal_bias="1.0"
-                            app:layout_constraintStart_toStartOf="parent"
-                            app:layout_constraintTop_toTopOf="parent"
-                            app:layout_constraintVertical_bias="0.0"/>
-
-                </FrameLayout>
-            </com.apachat.swipereveallayout.core.SwipeLayout>
-        </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-        <com.google.android.material.button.MaterialButton
-                android:id="@+id/quoteBookmarkButton"
-                style="@style/Button"
-                android:layout_width="48dp"
-                android:layout_height="48dp"
-                android:layout_marginBottom="16dp"
-                android:layout_marginEnd="16dp"
-                android:foreground="@drawable/ic_bookmark_empty"
-                android:foregroundTint="?attr/colorOnPrimary"
-                android:foregroundGravity="center"
-                android:padding="10dp"
-                android:textAllCaps="false"
-                android:textColor="@color/white"
-                android:textSize="16sp"
-                app:layout_constraintVertical_bias="1.0"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"/>
-
-        <com.google.android.material.button.MaterialButton
-                android:id="@+id/shareButton"
-                style="@style/Button"
-                android:layout_width="48dp"
-                android:layout_height="48dp"
-                android:layout_marginBottom="16dp"
-                android:layout_marginStart="16dp"
-                android:foreground="@drawable/ic_share"
-                android:foregroundGravity="center"
-                android:foregroundTint="?attr/colorOnPrimary"
-                android:padding="10dp"
-                android:textAllCaps="false"
-                android:textColor="@color/white"
-                android:textSize="16sp"
-                android:visibility="visible"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintVertical_bias="1.0"/>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <FrameLayout
+    <TextView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:id="@+id/collection_bottom_sheet"
-            app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior">
-        <include layout="@layout/collection_bottom_sheet" />
-    </FrameLayout>
+            android:text="@string/hello_blank_fragment"/>
 
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
+</FrameLayout>
