+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -6,10 +6,10 @@
             android:icon="@drawable/ic_home"
             android:title="@string/home"/>
     <item
-            android:id="@+id/navigation_search"
+            android:id="@+id/navigation_explore"
             android:enabled="true"
-            android:icon="@drawable/ic_search"
-            android:title="@string/search"/>
+            android:icon="@drawable/ic_explore"
+            android:title="@string/explore"/>
     <item
             android:id="@+id/navigation_create"
             android:enabled="true"
